move Footer component into Layout
